+++ src/App.js
@@ -1,6 +1,6 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 
-import Navbar from './Components/Navbar';
+import Navbar from './Components/Navbar/Navbar';
 import Teambuilder from './Pages/Teambuilder';
 import Teamviewer from './Pages/Teamviewer';
