+++ src/Components/Navbar/Navbar.js
@@ -1,4 +1,4 @@
-import PokeIcon from '../Icons/PokeBallIcon.svg'
+import PokeIcon from '../../Icons/PokeBallIcon.svg'
 
 const Navbar = () => {
     return (
